requirements.txt
streamlit>=1.0
pytest>=7.0
streamlit
streamlit-authenticator
pycryptodome
stegano
mysql-connector-python
google-api-python-client
google-auth-oauthlib
gspread
firebase_admin
hmac
st-cookie
google-api-python-client
google-auth-oauthlib